AudeX/requirements.txt
torch
transformers
datasets
gradio
Pillow
pytesseract
soundfile
git+https://github.com/huggingface/transformers.git
numpy
datasets
speechbrain
soundfile
librosa
